Move provider to package
